chore: remove event formatting for CampaignSignupMessage
